requirements.txt
asgiref==3.2.3
Click==7.0
cmake==3.13.3
dj-database-url==0.5.0
Django==3.0.3
dlib==19.19.0
face-recognition==1.2.3
face-recognition-models==0.3.0
gunicorn==20.0.4
numpy==1.18.1
opencv-python==4.2.0.32
packaging==20.1
Pillow==7.0.0
psycopg2==2.8.4
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
pytz==2019.3
pyzbar==0.1.8
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
